+++ visidata/__init__.py
@@ -3,6 +3,11 @@
 
 from .vd import __version__
 from .vd import *
+from .help import *
+from .async import *
+from .status_history import *
+from .zscroll import *
+
 from .addons.freqtbl import *
 from .addons.pyobj import *
 from .addons.metasheets import *
@@ -17,8 +22,5 @@ from .addons.zip import *
 from .addons.xlsx import *
 from .addons.hdf5 import *
 from .addons.sqlite import *
-#from .addons.gsheets import *
-#from .addons.blaze import *
-
 
 set_globals(globals())