+++ CHANGELOG.md
@@ -1,11 +1,27 @@
 # VisiData version history
 
-## v0.61
+## v0.91 (2017-06-28)
+- Make options automatically typed based on default
+- Documentation cleanups
+- Remove R command (set filetype on CLI)
+
+## v0.80
+- tour of screenshot.gif
+- regex transform now `*` (';' is still regex split)
+- Make regex search/select to work more like vim
+- Move several non-essential commands out of vd.py
+- change license of vd.py to MIT
+- vdtutor start
+- currency type with `$`; str type moved to `~`; remove type autodetect
+- www/ for landing page
+- move from .md to .rst for documentation
+
+## v0.61 (2017-06-12)
 - colorizers
 - `g[` and `g]` to sort by all key columns
 - `;` and `.` experimental regex commands
 
-## v0.59
+## v0.59 (2017-05-31)
 - pivot sheets with `W`
 - undo with `BACKSPACE` and replay with `ga`
 - dev guide and user guide